+++ src/index.tsx
@@ -1,24 +1,33 @@
 export { Element } from "./components/Element/Element";
 
+// LAYOUT  ////////////////////////////////////////////////////////////////////
 export { Row } from "./components/Row/Row";
 export { Portion } from "./components/Portion/Portion";
 
+// TEXT  //////////////////////////////////////////////////////////////////////
 export { Heading } from "./components/Typography/Heading";
 export { Text } from "./components/Typography/Text";
 
+// BREADCRUMBS  ///////////////////////////////////////////////////////////////
+export { BreadcrumbsWrapper } from "./components/Breadcrumbs/BreadcrumbsWrapper/BreadcrumbsWrapper";
+export { BreadcrumbItem } from "./components/Breadcrumbs/BreadcrumbItem/BreadcrumbItem";
+
+// BUTTON  ////////////////////////////////////////////////////////////////////
+export { Button } from "./components/Button/Button";
+
+// SIDEBAR  ///////////////////////////////////////////////////////////////////
 export { SidebarWrapper } from "./components/Sidebar/SidebarWrapper/SidebarWrapper";
 export { SidebarHeader } from "./components/Sidebar/SidebarHeader/SidebarHeader";
 export { SidebarItem } from "./components/Sidebar/SidebarItem/SidebarItem";
 export { SidebarItemIcon } from "./components/Sidebar/SidebarItemIcon/SidebarItemIcon";
 export { SidebarItemText } from "./components/Sidebar/SidebarItemText/SidebarItemText";
 export { SidebarFooter } from "./components/Sidebar/SidebarFooter/SidebarFooter";
-
 export { ContentWrapper } from "./components/ContentWrapper/ContentWrapper";
 
+// H-RULE  ////////////////////////////////////////////////////////////////////
 export { HRule } from "./components/HRule/HRule";
 
-export { Button } from "./components/Button/Button";
-
+// FORM  //////////////////////////////////////////////////////////////////////
 export { FormWrapper } from "./components/Form/FormWrapper/FormWrapper";
 export { FormItem } from "./components/Form/FormItem/FormItem";
 export { FormItemGroup } from "./components/Form/FormItemGroup/FormItemGroup";
@@ -32,34 +41,52 @@ export { RadioButton } from "./components/Form/RadioButton/RadioButton";
 export { CheckBox } from "./components/Form/CheckBox/CheckBox";
 export { Switch } from "./components/Form/CheckBox/Switch";
 
+// ACCORDION  /////////////////////////////////////////////////////////////////
 export { ExpandableContent } from "./components/ExpandableContent/ExpandableContent";
 
+// PROGRESS BAR  //////////////////////////////////////////////////////////////
 export { ProgressBar } from "./components/ProgressBar/ProgressBar";
 
+// CARD  //////////////////////////////////////////////////////////////////////
 export { Card } from "./components/Card/Card";
 
+// INFO PANEL  ////////////////////////////////////////////////////////////////
 export { InfoPanel } from "./components/InfoPanel/InfoPanel";
 
+// NOTIFICATIONS  /////////////////////////////////////////////////////////////
 export { NotificationsWrapper } from "./components/Notifications/NotificationsWrapper/NotificationsWrapper";
 export { NotificationItem } from "./components/Notifications/NotificationsItem/NotificationItem";
 
-export { BreadcrumbsWrapper } from "./components/Breadcrumbs/BreadcrumbsWrapper/BreadcrumbsWrapper";
-export { BreadcrumbItem } from "./components/Breadcrumbs/BreadcrumbItem/BreadcrumbItem";
 
+// TABLE  /////////////////////////////////////////////////////////////////////
 export { Table } from "./components/Table/Table";
 export { TablePagination } from "./components/Table/TablePagination/TablePagination";
 
+// CODE  //////////////////////////////////////////////////////////////////////
 export { CodeBlock } from "./components/CodeBlock/CodeBlock";
 
+// TABS  //////////////////////////////////////////////////////////////////////
 export { Tabs } from "./components/Tabs/Tabs";
-export { Spinner } from "./components/Spinner/Spinner";
-export { Callout } from "./components/Callout/Callout";
-export { Tag } from "./components/Tag/Tag";
-export { Tooltip } from "./components/Tooltip/Tooltip";
-export { ToastsWrapper } from "./components/Toast/ToastsWrapper/ToastsWrapper";
-export { Toast } from "./components/Toast/ToastItem/Toast";
 
-export { ThemeProvider } from "./components/ThemeProvider/ThemeProvider";
-export { ThemeType, ThemeProps, ColourPropTypes } from "./components/Element/constants";
+// SPINNER  ///////////////////////////////////////////////////////////////////
+export { Spinner } from "./components/Spinner/Spinner";
+
+// CALLOUT  ///////////////////////////////////////////////////////////////////
+export { Callout } from "./components/Callout/Callout";
+
+// TAG PILL  //////////////////////////////////////////////////////////////////
+export { Tag } from "./components/Tag/Tag";
+
+// TOOLTIP  ///////////////////////////////////////////////////////////////////
+export { Tooltip } from "./components/Tooltip/Tooltip";
+
+// TOAST  /////////////////////////////////////////////////////////////////////
+export { ToastsWrapper } from "./components/Toast/ToastsWrapper/ToastsWrapper";
+export { ToastItem } from "./components/Toast/ToastItem/ToastItem";
+
+// THEME  /////////////////////////////////////////////////////////////////////
 export { defaultColours } from "./styles/DefaultColours";
 export { FictoanTheme } from "./styles/theme";
+export { ThemeProvider } from "./components/ThemeProvider/ThemeProvider";
+export { ThemeType, ThemeProps, ColourPropTypes } from "./components/Element/constants";
+