+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### 0.37.1
+- Rename `Toast` to `ToastItem` for better consistency
+
 ### 0.37.0
 - Add `Toast` component
 
@@ -254,7 +257,7 @@ Breaking changes
 - Switch to useState & useEffect for merging theme
 
 ### 0.31.7
-- Fix Toast theme issue for setting `borderRadius`
+- Fix ToastItem theme issue for setting `borderRadius`
 
 ### 0.31.6
 - Fix broken builds (0.31.2 - 0.31.5)